mutt_realpath: toggle symlink resolution
This uses pre-existing non-symlinking impl of realpath() found in
`mutt_pretty_mailbox`. There is a call to realpath() for parent
directories in this code that needs to be removed. It means if you are
in a symlink and <change-dir>.. it will resolve. But I didn't want to
dirty this commit so I will do that next.

There is a concern about potentially overflowing string buffer in
`url_check_scheme`. That needs review.